finished verson
in the precious versions if I chose psychedelic option, it would show it's background color as psychedelic. There is nothing wrong with that, but if I chose White theme after psychedelic, the psychedelic background image would not change. So I had to make sure there will be no psychedelic background image in any other non-psychedelic themes